requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
attrs==25.3.0
audioop-lts==0.2.1
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
croniter==6.0.0
discord-py-interactions==5.14.0
discord-typings==0.9.0
discord.py==2.5.2
emoji==2.14.1
frozenlist==1.6.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
gspread==6.2.0
httplib2==0.22.0
idna==3.10
multidict==6.4.3
oauth2client==4.1.3
oauthlib==3.2.2
propcache==0.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
tomli==2.2.1
typing_extensions==4.13.2
urllib3==2.4.0
yarl==1.20.0
